+++ js/app/directives/onesc.js
@@ -4,18 +4,16 @@
  * This file is licensed under the Affero General Public License version 3 or
  * later. See the COPYING file.
  *
- * @author Volkan Gezer <volkangezer@gmail.com>
  * @author Pauli Järvinen <pauli.jarvinen@gmail.com>
- * @copyright 2014 Volkan Gezer
- * @copyright 2020 - 2023 Pauli Järvinen
+ * @copyright 2023 Pauli Järvinen
  *
  */
 
-angular.module('Music').directive('onEnter', ['$timeout', function ($timeout) {
+angular.module('Music').directive('onEsc', ['$timeout', function ($timeout) {
 	return function (scope, element, attrs) {
 		element.on('keydown keypress', (event) => {
-			if (event.which === 13) {
-				$timeout(() => scope.$eval(attrs.onEnter, {$event: event}));
+			if (event.which === 27) {
+				$timeout(() => scope.$eval(attrs.onEsc, {$event: event}));
 				event.preventDefault();
 			}
 		});